refactor(pkugin): abstruct proxy cache related function to kong.tools
