+++ kong/plugins/proxy-cache/handler.lua
@@ -23,6 +23,8 @@ local ngx_re_sub       = ngx.re.gsub
 local ngx_re_match     = ngx.re.match
 local parse_http_time  = ngx.parse_http_time
 local parse_mime_type  = mime_type.parse_mime_type
+local parse_directive_header = require("kong.tools.cache").parse_directive_header
+local calculate_resource_ttl = require("kong.tools.cache").calculate_resource_ttl
 
 
 local tab_new = require("table.new")
@@ -74,39 +76,6 @@ local function set_res_header(res, header, value, conf)
   end
 end
 
-local function parse_directive_header(h)
-  if not h then
-    return EMPTY
-  end
-
-  if type(h) == "table" then
-    h = concat(h, ", ")
-  end
-
-  local t    = {}
-  local res  = tab_new(3, 0)
-  local iter = ngx_re_gmatch(h, "([^,]+)", "oj")
-
-  local m = iter()
-  while m do
-    local _, err = ngx_re_match(m[0], [[^\s*([^=]+)(?:=(.+))?]],
-                                "oj", nil, res)
-    if err then
-      kong.log.err(err)
-    end
-
-    -- store the directive token as a numeric value if it looks like a number;
-    -- otherwise, store the string value. for directives without token, we just
-    -- set the key to true
-    t[lower(res[1])] = tonumber(res[2]) or res[2] or true
-
-    m = iter()
-  end
-
-  return t
-end
-
-
 local function req_cc()
   return parse_directive_header(ngx.var.http_cache_control)
 end
@@ -114,27 +83,6 @@ end
 
 local function res_cc()
   return parse_directive_header(ngx.var.sent_http_cache_control)
-end
-
-
-local function resource_ttl(res_cc)
-  local max_age = res_cc["s-maxage"] or res_cc["max-age"]
-
-  if not max_age then
-    local expires = ngx.var.sent_http_expires
-
-    -- if multiple Expires headers are present, last one wins
-    if type(expires) == "table" then
-      expires = expires[#expires]
-    end
-
-    local exp_time = parse_http_time(tostring(expires))
-    if exp_time then
-      max_age = exp_time - time()
-    end
-  end
-
-  return max_age and max(max_age, 0) or 0
 end
 
 
@@ -227,7 +175,7 @@ local function cacheable_response(conf, cc)
     return false
   end
 
-  if conf.cache_control and resource_ttl(cc) <= 0 then
+  if conf.cache_control and calculate_resource_ttl(cc) <= 0 then
     return false
   end
 
@@ -425,7 +373,7 @@ function ProxyCacheHandler:header_filter(conf)
   if cacheable_response(conf, cc) then
     -- TODO: should this use the kong.conf configured limit?
     proxy_cache.res_headers = resp_get_headers(0, true)
-    proxy_cache.res_ttl = conf.cache_control and resource_ttl(cc) or conf.cache_ttl
+    proxy_cache.res_ttl = conf.cache_control and calculate_resource_ttl(cc) or conf.cache_ttl
 
   else
     set_header(conf, "X-Cache-Status", "Bypass")
